requirements.txt
torch
torchvision
numpy
Pillow
gradio